improve(sign-in-with-email): moves it into its own dedicated file before fragmenting it by context (provider, register / login etc)
